requirements.txt
fpd
gradio
langgraph
smolagents
transformers
pandas
matplotlib
seaborn
textblob
requests
python-dotenv